% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continents.R
\docType{data}
\name{continents}
\alias{continents}
\title{SDB table "continents"}
\format{
\subsection{\code{continents}}{

A data frame with 6 rows and 2 columns:
\describe{
\item{continent_id}{record sequence number}
\item{continent_name}{Continent Name}
}
}
}
\source{
membership <- tbl(con, "continents") |> collect()
}
\usage{
continents
}
\description{
A convenient copy of SDB table "continents"
}
\keyword{datasets}
