% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_init.R
\name{db_init}
\alias{db_init}
\title{Open a connection to the SDB}
\usage{
db_init(db = c("sdb", "network"))
}
\arguments{
\item{db}{which of 2 databases to query}
}
\value{
con object
}
\description{
\code{sdb_init} runs 2 queries to insure that downloads come back in utf8 format
}
