% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_type.R
\docType{data}
\name{location_type}
\alias{location_type}
\title{SDB table "location_type"}
\format{
\subsection{\code{location_type}}{

A data frame with 8 rows and 2 columns:
\describe{
\item{id}{Location Code}
\item{location_type}{Location Type}
}
}
}
\source{
location_type <- tbl(con, "location_type")
|> collect() |> rename(location_type = name)
}
\usage{
location_type
}
\description{
A convenient copy of SDB table "location_type"
}
\keyword{datasets}
