% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_cat_center_size_count.R
\name{mrr_cat_center_size_count}
\alias{mrr_cat_center_size_count}
\title{Categorize values into ordered size categories.
Center size is categorized into five categories based on the 2025
distribution of the number of members in the center:
Very Small ( < 10),
Small (11 - 20),
Medium (21 - 50),
Large (51 - 100),
Very Large ( > 101)}
\usage{
mrr_cat_center_size_count(
  df,
  n_members_var,
  center_size_category = "center_size"
)
}
\arguments{
\item{df}{A data frame}

\item{n_members_var}{Variable containing numeric values to categorize}

\item{center_size_category}{Name for the new categorical variable. Default: "center_size"}
}
\value{
Data frame with added ordered factor column for size categories
}
\description{
Categorize values into ordered size categories.
Center size is categorized into five categories based on the 2025
distribution of the number of members in the center:
Very Small ( < 10),
Small (11 - 20),
Medium (21 - 50),
Large (51 - 100),
Very Large ( > 101)
}
\examples{
\dontrun{

df <- mrr_cat_center_size_count(
  base_center_data,
  center_members_now,
  center_size_category
)

df |>
  ggplot(aes(center_size_category, log(center_members_now))) +
  geom_boxplot() +
  coord_flip()
}
}
