% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_cat_center_size_members.R
\name{mrr_cat_center_size_members}
\alias{mrr_cat_center_size_members}
\title{Categorize centers by member count into ordered size categories.
Center size is categorized into five categories based on the 2025
distribution of the number of members in the center:
Very Small ( < 10),
Small (11 - 20),
Medium (21 - 50),
Large (51 - 100),
Very Large ( > 101)}
\usage{
mrr_cat_center_size_members(
  df,
  center_var = center,
  membertype_var = membertype,
  center_size_var = center_size
)
}
\arguments{
\item{df}{A data frame with individual member data}

\item{center_var}{Column identifying centers}

\item{membertype_var}{Column identifying member types}

\item{center_size_var}{Name for output size category column}
}
\value{
Data frame with added ordered factor column for center size categories
}
\description{
Categorize centers by member count into ordered size categories.
Center size is categorized into five categories based on the 2025
distribution of the number of members in the center:
Very Small ( < 10),
Small (11 - 20),
Medium (21 - 50),
Large (51 - 100),
Very Large ( > 101)
}
\details{
Preserves existing group structure of input data frame
}
\examples{
\dontrun{
center_size_df <- mrr_cat_center_size_members(
  df = member_base_data,
  center_var = center,
  membertype_var = membertype,
  center_size_var = center_size
)

center_size_df |> janitor::tabyl(center_size)
}
}
