% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_color_levels.R
\name{mrr_color_steps}
\alias{mrr_color_steps}
\title{Generate a color palette with a specified number of steps}
\usage{
mrr_color_steps(color_number, levels, blend_number = 4)
}
\arguments{
\item{color_number}{An integer specifying the base color palette to use}

\item{levels}{An integer specifying the number of steps to include in the final palette}

\item{blend_number}{An integer specifying the color palette to blend with the base palette}
}
\value{
A vector of colors with the specified number of steps
}
\description{
Generate a color palette with a specified number of steps
}
\examples{
mrr_color_steps(5, 5, 4)

}
