% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_lolly.R
\name{mrr_comparison_lolly}
\alias{mrr_comparison_lolly}
\title{single and multiple lollypop plots}
\usage{
mrr_comparison_lolly(df)
}
\arguments{
\item{df}{a data frame containing at least 3 variables: c_pct, c_pct_then, and var_label}
}
\value{
plot
}
\description{
\code{mrr_comparison_lolly} creates a lollypop plot comparing two variables
}
