% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_count}
\alias{mrr_count}
\title{count the frequency of a single variable}
\usage{
mrr_count(c_df, c_var, c_now = TRUE)
}
\arguments{
\item{c_df}{The dataset to be counted}

\item{c_var}{The variable with values to be counted}
}
\value{
df_count
}
\description{
\code{mrr_count} returns a data frame with the counts and percent of total
}
