% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_count_multiple_vars}
\alias{mrr_count_multiple_vars}
\title{count the frequency of a single variable}
\usage{
mrr_count_multiple_vars(c_df, c_start_var, c_end_var, c_df_name)
}
\arguments{
\item{c_df}{The dataset to be counted}

\item{c_start_var}{The first variable in a sequence to be counted}

\item{c_end_var}{The variable in a sequence to be counted}
}
\value{
df_count
}
\description{
\code{mrr_count_multiple_vars} returns a data frame with the counts and percent of total for a series of variables

All of the variables in the series must have the same factor structure
}
