% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_count_single}
\alias{mrr_count_single}
\title{count the frequency of a single variable}
\usage{
mrr_count_single(c_df, c_var, c_now = TRUE)
}
\arguments{
\item{c_df}{The dataset to be counted}

\item{c_var}{The variable with values to be counted}

\item{c_now}{A flag that appends "then" to putput variables when set to "FALSE"}
}
\value{
df_count
}
\description{
\code{mrr_count_single} returns a data frame with the counts and percent of total
}
\details{
Here is an example of using 'mrr_count_single'
count_df_now <- m23 |>  mrr_count(c_var = "relate_to_group_or_center") |>
mutate(c_count_var = fct_relabel(c_count_var, word, 1))

params <- m23 |>  mrr_get_dd(c_varname =  "relate_to_group_or_center")
params$count_df23 <- count_df_now

count_df_then <- m22 |>  mrr_count(c_var = "yes_relate_to_a_local_center", c_now = FALSE ) |>
mutate(c_count_var_then = fct_relabel(c_count_var_then, word, 1))

params$count_df <- bind_cols(count_df_now, count_df_then)
}
