% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_count_var_sequence}
\alias{mrr_count_var_sequence}
\title{count the frequency of a sequence of variables}
\usage{
mrr_count_var_sequence(c_df, c_start_var, c_end_var, c_df_name)
}
\arguments{
\item{c_df}{The dataset to be counted}

\item{c_start_var}{The first variable in a sequence to be counted}

\item{c_end_var}{The variable in a sequence to be counted}

\item{c_df_name}{the survey id, for chained calls where 'c_df' is not \emph{true}}
}
\value{
df_count
}
\description{
\code{mrr_count_var_sequence} returns a data frame with the counts and percent of total one or more variables
}
\details{
For the params list, use:

params <- mrr_get_dd(m23, "people_in_center_can_give_help_emotional_support")

For the counts, use:
params$count_df <- m23 |>
mrr_count_var_sequence(
c_start_var = people_in_center_can_give_help_emotional_support,
c_end_var = people_in_center_can_give_advice_about_practice_etc,
"m23")
}
\section{may need to rename variables afterwards:}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  if (c_now) \{
    names(df_count) <- c("c_var_name", "c_n", "c_pct" )
    \} else\{
      names(df_count) <- c("c_count_var_then", "c_n_then", "c_pct_then")
}\if{html}{\out{</div>}}
}

