% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_get_classifying_vars}
\alias{mrr_get_classifying_vars}
\title{Get a data frame with respondent ID and specified classifying variables from a given data frame}
\usage{
mrr_get_classifying_vars(df, classify_var_list)
}
\arguments{
\item{df}{Data frame to extract the variables from}

\item{classify_var_list}{List of classifying variables to extract}
}
\value{
Data frame with respondent ID and specified classifying variables
}
\description{
Get a data frame with respondent ID and specified classifying variables from a given data frame
}
\examples{
df <- data.frame(respondent_id = c(1, 2, 3),
                  age = c(20, 30, 25),
                  gender = c('M', 'F', 'M'))

mrr_get_classifying_vars(df, c('age', 'gender'))

}
