% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_get_dd}
\alias{mrr_get_dd}
\title{count the frequency of a single variable and join with metadata}
\usage{
mrr_get_dd(c_dfname, c_varname)
}
\arguments{
\item{c_dfname}{documented in the dd_all data frame}

\item{c_varname}{The variable to be documented}
}
\value{
params
}
\description{
\code{mrr_get_dd} returns a list that contains metadata
}
