% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_job_doc.R
\name{mrr_get_last_updated_source}
\alias{mrr_get_last_updated_source}
\title{Get the most recently modified file with specified extensions}
\usage{
mrr_get_last_updated_source()
}
\value{
A character string containing the path of the most recently modified file
}
\description{
This function retrieves the most recently modified file with extensions
specified by the regular expression pattern. The function searches for
files with extensions 'Rmd', 'Qmd', and 'R'. It returns the path of the
most recently modified file.
}
