% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_job_doc.R
\name{mrr_job_doc}
\alias{mrr_job_doc}
\title{Generates a tibble with information about the job}
\usage{
mrr_job_doc()
}
\value{
A tibble with a single column and three rows:
file_doc, run_date, and r_version.
}
\description{
This function generates a tibble with information about the job, including
the file name, run date, and R version.
}
