% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_meta.R
\name{mrr_meta}
\alias{mrr_meta}
\title{Convenience function to get data dictionary info}
\usage{
mrr_meta(c_df_name, c_varname)
}
\arguments{
\item{c_df_name}{documented in the dd_all data frame}

\item{c_varname}{The variable to be documented}
}
\value{
meta_list
}
\description{
\code{mrr_meta} returns a list that contains metadata
}
