% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_lolly.R
\name{mrr_single_lolly}
\alias{mrr_single_lolly}
\title{single and multiple lollypop plots}
\usage{
mrr_single_lolly(mrr_params)
}
\arguments{
\item{mrr_params}{a list containing metadata elements and a data frame named count$df}
}
\value{
plot
}
\description{
\code{mrr_single_lolly} creates a lollypop plot for a single variable
}
