% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_waffles.R
\name{mrr_waffle}
\alias{mrr_waffle}
\title{Creates a waffle plot with perceived membership data}
\usage{
mrr_waffle(df, fill_var, values_var)
}
\arguments{
\item{df}{A dataframe with perceived membership data.}

\item{fill_var}{A factor representing the classification categories in the dataframe}

\item{values_var}{A numeric vector with the percentage of people belonging to each category}
}
\value{
A waffle plot of the data classified in the dataframe
}
\description{
This function creates a waffle plot of perceived membership data using ggplot2 and geom_waffle packages.
}
