% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_to_titles.R
\name{names_to_titles}
\alias{names_to_titles}
\title{Replaces underscores with blanks and capitalizes first letter of each word}
\usage{
names(df) <- names_to_titles(df)
}
\arguments{
\item{df}{data frame where column names will be modified}
}
\value{
modified column names with underscores replaced by blanks
}
\description{
Replaces underscores with blanks and capitalizes first letter of each word
}
\examples{
\dontrun{
# names(df) <- names_to_titles(df)
}

}
