% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourmirror-package.R
\docType{package}
\name{ourmirror-package}
\alias{ourmirror}
\alias{ourmirror-package}
\title{ourmirror: Provides a default theme and other functions for the Societal Mirror}
\description{
Sets standards for ggplot, accesses the SDB, and provides other necessary functions in use across all SGS-IT work.
}
\author{
\strong{Maintainer}: John Smith \email{john.smith@shambhalaglobal.org} [contributor]

}
\keyword{internal}
