% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_palettes.R
\name{scale_colour_mirror_d}
\alias{scale_colour_mirror_d}
\alias{scale_color_mirror_d}
\title{Discrete brand colour scale for ggplot2}
\usage{
scale_colour_mirror_d(...)

scale_color_mirror_d(...)
}
\arguments{
\item{...}{Passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 colour scale.
}
\description{
Discrete brand colour scale for ggplot2
}
