% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_palettes.R
\name{scale_fill_mirror_d}
\alias{scale_fill_mirror_d}
\title{Discrete brand fill scale for ggplot2}
\usage{
scale_fill_mirror_d(...)
}
\arguments{
\item{...}{Passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 fill scale.
}
\description{
Discrete brand fill scale for ggplot2
}
