% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_dots.R
\name{strip_dots}
\alias{strip_dots}
\title{Remove Dots from Factor Levels or Character Strings}
\usage{
strip_dots(x)
}
\arguments{
\item{x}{A factor, character vector, or other variable. Factors and characters
will have dots removed; other types are returned unchanged.}
}
\value{
A factor or character vector with dots removed, or the original value
if the input is neither factor nor character.
}
\description{
Removes all dots/periods from the levels of a factor variable or from
character strings using string replacement. This is useful for cleaning
factor levels that contain dots from data imports.
}
\examples{
# With factors
x <- factor(c("a.b", "c.d", "e.f"))
strip_dots(x)
# Returns factor with levels "ab", "cd", "ef"

# With characters
y <- c("hello.world", "foo.bar")
strip_dots(y)
# Returns c("helloworld", "foobar")

# With other types (returned unchanged)
strip_dots(42)
# Returns 42

}
