% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_html_tags.R
\name{strip_html_tags}
\alias{strip_html_tags}
\title{Strip HTML tags and decode entities from text}
\usage{
strip_html_tags(html_string)
}
\arguments{
\item{html_string}{A character vector containing HTML strings to clean}
}
\value{
A character vector with HTML tags removed and entities decoded
}
\description{
Removes HTML tags and decodes common HTML entities from character vectors.
Handles NA and empty values gracefully.
}
\examples{
strip_html_tags("<p>Hello &amp; welcome</p>")
strip_html_tags(c("<b>Bold</b>", NA, ""))

}
