% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_underscore.R
\name{strip_underscore}
\alias{strip_underscore}
\title{Replaces underscores with blanks for readability in a google spreadsheet}
\usage{
strip_underscore(df)
}
\arguments{
\item{df}{data frame where column names will be modified}
}
\value{
modified column names with underscores replaced by blanks
}
\description{
Replaces underscores with blanks for readability in a google spreadsheet
}
\examples{
\dontrun{
# names(df) <- strip_underscore(df)
}

}
