% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{SDB table "countries"}
\format{
\subsection{\code{countries}}{

A data frame with 241 rows and 6 columns:
\describe{
\item{id}{record sequence number}
\item{country}{country name}
\item{country_code}{numeric country code}
\item{continent}{continent code number}
\item{top_level_domain}{Internet domain}
\item{currency}{Currency}
}
}
}
\source{
countries <- tbl(con, "countries") |> collect()
}
\usage{
countries
}
\description{
A convenient copy of SDB table "countries"
}
\keyword{datasets}
