% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_type.R
\docType{data}
\name{member_type}
\alias{member_type}
\title{SDB table "member_type"}
\format{
\subsection{\code{member_type}}{

A data frame with 6 rows andmember_type2 columns:
\describe{
\item{member_type_id}{Member Type ID Number}
\item{member_type_name}{Member Type Label}
}
}
}
\source{
membership <- tbl(con, "membership") |> collect()
}
\usage{
member_type
}
\description{
A convenient copy of SDB table "membership"
}
\keyword{datasets}
