% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_theme.R
\name{mirror_theme}
\alias{mirror_theme}
\title{Create a ggplot2 theme}
\usage{
mirror_theme(font = "Helvetica", base_size = 14)
}
\arguments{
\item{font}{default parameter}

\item{base_size}{default font size}
}
\value{
ggplot2::theme
}
\description{
\code{mirror_theme} creates a ggplot2 theme
}
