% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_factors.R
\name{mrr_append_factor_pct}
\alias{mrr_append_factor_pct}
\title{Append Percentages to Categorical Variable Labels}
\usage{
mrr_append_factor_pct(factor_var)
}
\arguments{
\item{factor_var}{A categorical variable to transform}
}
\value{
A named character vector with original labels appended with percentages
}
\description{
This function appends the percentage of observations in each category
to the label of a categorial variable.
}
\examples{
x <- c('A', 'A', 'B', 'C', 'C', 'C')
mrr_append_factor_pct(x)

}
\author{
John David Smith
}
