% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_color_levels.R
\name{mrr_color_both_ways}
\alias{mrr_color_both_ways}
\title{Generate a color palette with a specified number of steps}
\usage{
mrr_color_both_ways(color_number, levels)
}
\arguments{
\item{color_number}{An integer specifying the base color palette to use}

\item{levels}{An integer specifying the number of steps to include in the final palette}
}
\value{
A vector of colors with the specified number of steps
}
\description{
Generate a color palette with a specified number of steps
}
