% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_lolly.R
\name{mrr_comparison_lolly}
\alias{mrr_comparison_lolly}
\title{single and multiple lollypop plots}
\usage{
mrr_comparison_lolly(mrr_params)
}
\arguments{
\item{mrr_params}{a list containing metadata elements and a data frame named count$df}
}
\value{
plot
}
\description{
\code{mrr_comparison_lolly} creates a lollypop plot comparing two variables
}
