% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_count_var_list}
\alias{mrr_count_var_list}
\title{count the frequency of a single variable}
\usage{
mrr_count_var_list(c_df, c_var_list, c_df_name)
}
\arguments{
\item{c_df}{The dataset to be counted}

\item{c_var_list}{The list of variables to be counted}

\item{c_df_name}{the survey id, for chained calls where 'c_df' is not \emph{true}}
}
\value{
df_count
}
\description{
\code{mrr_count_var_list} returns a data frame with the counts and percent of total for a list of variables

The variables must have the same factor structure
}
