% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_count_var_list}
\alias{mrr_count_var_list}
\title{count the frequency of a single variable}
\usage{
mrr_count_var_list(c_df, c_var_list)
}
\arguments{
\item{c_df}{The dataset to be counted}

\item{c_var_list}{The list of variables to be counted}
}
\value{
df_count
}
\description{
\code{mrr_count_var_list} returns a data frame with the counts and percent of total for a list of variables

The variables have to have the same factor structure
}
