% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_count.R
\name{mrr_get_multi_vars}
\alias{mrr_get_multi_vars}
\title{Get Multiple Response Variables}
\usage{
mrr_get_multi_vars(df, response_var_list)
}
\arguments{
\item{df}{dataframe to extract from}

\item{response_var_list}{vector of variable names to extract}
}
\value{
Returns a data frame containing the respondent_id and all variables specified in response_var_list pivoted with pivot_longer
}
\description{
Extracts a data frame with multiple response variables from a given data frame
}
