% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_gmail.R
\name{mrr_gmail}
\alias{mrr_gmail}
\title{Send email using Gmail API}
\usage{
mrr_gmail(sender, recipient, title, text)
}
\arguments{
\item{sender}{The email address of the sender.}

\item{recipient}{The email address of the recipient.}

\item{title}{The subject of the email.}

\item{text}{The content of the email.}
}
\description{
This function allows you to send an email using the Gmail API.
}
\examples{
# example code

\dontrun{
mrr_gmail("sender@gmail.com", "recipient@gmail.com",
"Hello", "This is the body of the email.")
}
}
