% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_lolly.R
\name{mrr_single_lolly}
\alias{mrr_single_lolly}
\title{single and multiple lollypop plots}
\usage{
mrr_single_lolly(count_df)
}
\arguments{
\item{count_df}{a data frame with variables var_response, c_pct,}
}
\value{
plot
}
\description{
\code{mrr_single_lolly} creates a lollypop plot for a single variable
}
