% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrr_color_levels.R
\name{shambhala_palette_function}
\alias{shambhala_palette_function}
\title{Shambhala color palette function}
\usage{
shambhala_palette_function()
}
\value{
A character vector containing the Shambhala color palette.
}
\description{
This function creates a color palette for Societal Mirror website.
}
\examples{
shambhala_palette()

}
